+++ src/__tests__/userDataProviderPermissions.int.test.ts
@@ -3,6 +3,8 @@ import { app } from '../app';
 import {
 	createOrUpdateDataProvider,
 	createOrUpdateUserDataProviderPermission,
+	loadUserDataProviderPermission,
+	removeUserDataProviderPermission,
 	loadSystemUser,
 } from '../database';
 import { expectTimestamp, loadTestUser } from '../test/utils';
@@ -11,6 +13,7 @@ import {
 	mockJwtWithAdminRole as authHeaderWithAdminRole,
 } from '../test/mockJwt';
 import { keycloakIdToString, Permission } from '../types';
+import { NotFoundError } from '../errors';
 
 describe('/users/dataProviders/:dataProviderShortcode/permissions/:permission', () => {
 	describe('PUT /', () => {
@@ -161,6 +164,202 @@ describe('/users/dataProviders/:dataProviderShortcode/permissions/:permission', 
 				permission: Permission.MANAGE,
 				userKeycloakUserId: user.keycloakUserId,
 			});
+		});
+	});
+
+	describe('DELETE /', () => {
+		it('returns 401 if the request lacks authentication', async () => {
+			const user = await loadTestUser();
+			const dataProvider = await createOrUpdateDataProvider({
+				shortCode: 'ExampleInc',
+				name: 'Example Inc.',
+				keycloakOrganizationId: null,
+			});
+			await request(app)
+				.delete(
+					`/users/${keycloakIdToString(user.keycloakUserId)}/dataProviders/${dataProvider.shortCode}/permissions/${Permission.MANAGE}`,
+				)
+				.send()
+				.expect(401);
+		});
+
+		it('returns 401 if the authenticated user lacks permission', async () => {
+			const user = await loadTestUser();
+			const dataProvider = await createOrUpdateDataProvider({
+				shortCode: 'ExampleInc',
+				name: 'Example Inc.',
+				keycloakOrganizationId: null,
+			});
+			await request(app)
+				.delete(
+					`/users/${keycloakIdToString(user.keycloakUserId)}/dataProviders/${dataProvider.shortCode}/permissions/${Permission.MANAGE}`,
+				)
+				.set(authHeader)
+				.send()
+				.expect(401);
+		});
+
+		it('returns 400 if the userId is not a valid keycloak user ID', async () => {
+			await request(app)
+				.delete(
+					`/users/notaguid/dataProviders/1/permissions/${Permission.MANAGE}`,
+				)
+				.set(authHeaderWithAdminRole)
+				.send()
+				.expect(400);
+		});
+
+		it('returns 400 if the data provider shortCode is not a valid shortCode', async () => {
+			const user = await loadTestUser();
+			await request(app)
+				.delete(
+					`/users/${keycloakIdToString(user.keycloakUserId)}/dataProviders/not a valid short code/permissions/${Permission.MANAGE}`,
+				)
+				.set(authHeaderWithAdminRole)
+				.send()
+				.expect(400);
+		});
+
+		it('returns 400 if the permission is not a valid permission', async () => {
+			const user = await loadTestUser();
+			await request(app)
+				.delete(
+					`/users/${keycloakIdToString(user.keycloakUserId)}/dataProviders/ExampleInc/permissions/notAPermission`,
+				)
+				.set(authHeaderWithAdminRole)
+				.send()
+				.expect(400);
+		});
+
+		it('returns 404 if the permission does not exist', async () => {
+			const user = await loadTestUser();
+			const dataProvider = await createOrUpdateDataProvider({
+				shortCode: 'ExampleInc',
+				name: 'Example Inc.',
+				keycloakOrganizationId: null,
+			});
+			await request(app)
+				.delete(
+					`/users/${keycloakIdToString(user.keycloakUserId)}/dataProviders/${dataProvider.shortCode}/permissions/${Permission.MANAGE}`,
+				)
+				.set(authHeaderWithAdminRole)
+				.send()
+				.expect(404);
+		});
+
+		it('returns 404 if the permission had existed and previously been deleted', async () => {
+			const user = await loadTestUser();
+			const dataProvider = await createOrUpdateDataProvider({
+				shortCode: 'ExampleInc',
+				name: 'Example Inc.',
+				keycloakOrganizationId: null,
+			});
+			await createOrUpdateUserDataProviderPermission({
+				userKeycloakUserId: user.keycloakUserId,
+				dataProviderShortCode: dataProvider.shortCode,
+				permission: Permission.EDIT,
+				createdBy: user.keycloakUserId,
+			});
+			await removeUserDataProviderPermission(
+				user.keycloakUserId,
+				dataProvider.shortCode,
+				Permission.EDIT,
+			);
+			await request(app)
+				.delete(
+					`/users/${keycloakIdToString(user.keycloakUserId)}/dataProviders/${dataProvider.shortCode}/permissions/${Permission.EDIT}`,
+				)
+				.set(authHeaderWithAdminRole)
+				.send()
+				.expect(404);
+		});
+
+		it('deletes the user data provider permission when the user has administrative credentials', async () => {
+			const user = await loadTestUser();
+			const dataProvider = await createOrUpdateDataProvider({
+				shortCode: 'ExampleInc',
+				name: 'Example Inc.',
+				keycloakOrganizationId: null,
+			});
+			await createOrUpdateUserDataProviderPermission({
+				userKeycloakUserId: user.keycloakUserId,
+				dataProviderShortCode: dataProvider.shortCode,
+				permission: Permission.EDIT,
+				createdBy: user.keycloakUserId,
+			});
+			const permission = await loadUserDataProviderPermission(
+				user.keycloakUserId,
+				dataProvider.shortCode,
+				Permission.EDIT,
+			);
+			expect(permission).toEqual({
+				dataProviderShortCode: dataProvider.shortCode,
+				createdAt: expectTimestamp,
+				createdBy: user.keycloakUserId,
+				permission: Permission.EDIT,
+				userKeycloakUserId: user.keycloakUserId,
+			});
+			await request(app)
+				.delete(
+					`/users/${keycloakIdToString(user.keycloakUserId)}/dataProviders/${dataProvider.shortCode}/permissions/${Permission.EDIT}`,
+				)
+				.set(authHeaderWithAdminRole)
+				.send()
+				.expect(204);
+			await expect(
+				loadUserDataProviderPermission(
+					user.keycloakUserId,
+					dataProvider.shortCode,
+					Permission.EDIT,
+				),
+			).rejects.toThrow(NotFoundError);
+		});
+
+		it('deletes the user data provider permission when the user has permission to manage the data provider', async () => {
+			const user = await loadTestUser();
+			const dataProvider = await createOrUpdateDataProvider({
+				shortCode: 'ExampleInc',
+				name: 'Example Inc.',
+				keycloakOrganizationId: null,
+			});
+			await createOrUpdateUserDataProviderPermission({
+				userKeycloakUserId: user.keycloakUserId,
+				dataProviderShortCode: dataProvider.shortCode,
+				permission: Permission.MANAGE,
+				createdBy: user.keycloakUserId,
+			});
+			await createOrUpdateUserDataProviderPermission({
+				userKeycloakUserId: user.keycloakUserId,
+				dataProviderShortCode: dataProvider.shortCode,
+				permission: Permission.EDIT,
+				createdBy: user.keycloakUserId,
+			});
+			const permission = await loadUserDataProviderPermission(
+				user.keycloakUserId,
+				dataProvider.shortCode,
+				Permission.EDIT,
+			);
+			expect(permission).toEqual({
+				dataProviderShortCode: dataProvider.shortCode,
+				createdAt: expectTimestamp,
+				createdBy: user.keycloakUserId,
+				permission: Permission.EDIT,
+				userKeycloakUserId: user.keycloakUserId,
+			});
+			await request(app)
+				.delete(
+					`/users/${keycloakIdToString(user.keycloakUserId)}/dataProviders/${dataProvider.shortCode}/permissions/${Permission.EDIT}`,
+				)
+				.set(authHeader)
+				.send()
+				.expect(204);
+			await expect(
+				loadUserDataProviderPermission(
+					user.keycloakUserId,
+					dataProvider.shortCode,
+					Permission.EDIT,
+				),
+			).rejects.toThrow(NotFoundError);
 		});
 	});
 });