+++ src/handlers/userDataProviderPermissionsHandlers.ts
@@ -1,4 +1,8 @@
-import { createOrUpdateUserDataProviderPermission } from '../database';
+import {
+	assertUserDataProviderPermissionExists,
+	createOrUpdateUserDataProviderPermission,
+	removeUserDataProviderPermission,
+} from '../database';
 import {
 	isAuthContext,
 	isId,
@@ -14,6 +18,61 @@ import {
 	InputValidationError,
 } from '../errors';
 import type { Request, Response, NextFunction } from 'express';
+
+const deleteUserDataProviderPermission = (
+	req: Request,
+	res: Response,
+	next: NextFunction,
+): void => {
+	const { userKeycloakUserId, dataProviderShortCode, permission } = req.params;
+	if (!isKeycloakId(userKeycloakUserId)) {
+		next(
+			new InputValidationError(
+				'Invalid userKeycloakUserId parameter.',
+				isKeycloakId.errors ?? [],
+			),
+		);
+		return;
+	}
+	if (!isShortCode(dataProviderShortCode)) {
+		next(
+			new InputValidationError(
+				'Invalid dataProviderShortCode parameter.',
+				isId.errors ?? [],
+			),
+		);
+		return;
+	}
+	if (!isPermission(permission)) {
+		next(
+			new InputValidationError(
+				'Invalid permission parameter.',
+				isPermission.errors ?? [],
+			),
+		);
+		return;
+	}
+
+	(async () => {
+		await assertUserDataProviderPermissionExists(
+			userKeycloakUserId,
+			dataProviderShortCode,
+			permission,
+		);
+		await removeUserDataProviderPermission(
+			userKeycloakUserId,
+			dataProviderShortCode,
+			permission,
+		);
+		res.status(204).contentType('application/json').send();
+	})().catch((error: unknown) => {
+		if (isTinyPgErrorWithQueryContext(error)) {
+			next(new DatabaseError('Error deleting item.', error));
+			return;
+		}
+		next(error);
+	});
+};
 
 const putUserDataProviderPermission = (
 	req: Request,
@@ -85,6 +144,7 @@ const putUserDataProviderPermission = (
 };
 
 const userDataProviderPermissionsHandlers = {
+	deleteUserDataProviderPermission,
 	putUserDataProviderPermission,
 };